test with 2 by Abdulbosit

+++ src/main/java/uz/pdp/backend/olxapp/service/CategoryServiceImpl.java
@@ -2,6 +2,7 @@ package uz.pdp.backend.olxapp.service;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,6 +17,7 @@ import uz.pdp.backend.olxapp.repository.CategoryRepository;
 import java.util.Collections;
 import java.util.List;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class CategoryServiceImpl implements CategoryService {
@@ -25,6 +27,7 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public List<CategoryDTO> getAllCategories() {
 
+        log.info("Getting all categories");
         List<Category> categories = categoryRepository.getCategory();
         return categories.stream().map(categoryMapper::toDto).toList();
 
@@ -85,6 +88,10 @@ public class CategoryServiceImpl implements CategoryService {
 
         if (!category.getChildren().isEmpty()) {
             throw new EntityNotFoundException("Can't delete category with child categories!", HttpStatus.CONFLICT);
+        }
+
+        if (!category.getProducts().isEmpty()) {
+            throw new EntityNotFoundException("Can't delete category with products!", HttpStatus.CONFLICT);
         }
 
         categoryRepository.delete(category);
